change measure unit type from int to float64; make test

+++ calc/pkg/calc_messages.go
@@ -17,7 +17,7 @@ type BusinessRequest struct {
 }
 
 type BusinessResponse struct {
-	Distance int    `json:"distance"`
-	Duration int    `json:"duration"`
-	Err      string `json:"err,omitempty"`
+	Distance float64 `json:"distance"`
+	Duration float64 `json:"duration"`
+	Err      string  `json:"err,omitempty"`
 }
